dom/html/test/test_bug787134.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=787134
-->
<head>
  <title>Test for Bug 787134</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="reflect.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=787134">Mozilla Bug 787134</a>
<p id="display"></p>
<p><a id="link-test1" href="example link">example link</a></p>
<pre id="test">
<script>
    var div = document.createElement('div');
    div.innerHTML = '<a href=#></a>';
    var a = div.firstChild;
    ok(a.matches(':link'), "Should match a link not in a document");
    is(div.querySelector(':link'), a, "Should find a link not in a document");
    a = document.querySelector('#link-test1');
    ok(a.matches(':link'), "Should match a link in a document with an invalid URL");
</script>
</pre>
</body>
</html>
